Backend/requirements.txt
Django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.4.0
django-ratelimit==4.1.0